fix: normalize generated MDX translations
