+++ .github/workflows/content-generation.yml
@@ -55,7 +55,11 @@ jobs:
         run: |
           node write-heading-ids.mjs
         continue-on-error: true
-      
+
+      - name: Normalize translated MDX
+        run: |
+          node normalize-translated-mdx.mjs i18n
+
       - name: Build
         run: pnpm build
         continue-on-error: true
@@ -102,4 +106,3 @@ jobs:
             bot
           draft: false
           delete-branch: true
-
