+++ i18n/es/docusaurus-plugin-content-docs/current/user-management/manage-users.mdx
@@ -18,7 +18,7 @@ La tabla consta de tres columnas:
 
 Admite mapeo de palabras clave para [`name`](/user-management/user-data#name), [`id`](/user-management/user-data#id), [`username`](/user-management/user-data#username), [`primary-phone`](/user-management/user-data#primary_phone), [`primary-email`](/user-management/user-data#primary_email).
 
-### Añadir usuarios{add-users} \{#add-users}
+### Añadir usuarios \{#add-users}
 
 Usando la Consola, los desarrolladores pueden crear nuevas cuentas para los usuarios finales. Para hacerlo, haz clic en el botón "Añadir usuario" en la esquina superior derecha de la pantalla.
