+++ i18n/ja/docusaurus-plugin-content-docs/current/customization/bring-your-ui/README.mdx
@@ -52,7 +52,8 @@ Logto プロジェクト全体は [pnpm モノレポ](https://pnpm.io/workspaces
 
 1. <CloudLink to="/sign-in-experience/branding">
      コンソール > サインイン & アカウント > ブランディング > 独自 UI の持ち込み
-   </CloudLink> に移動します。
+   </CloudLink>
+   に移動します。
 2. 前の手順で作成した zip ファイルをクリックして選択、またはドラッグ＆ドロップすると、アップロードが自動的に開始されます。
 3. アップロードが完了したら変更を保存してください。カスタム UI が即座に提供されます。
 4. 独自のサインイン UI を使用している場合、「サインインプレビュー」ウィンドウは無効になります。ただし、 [ライブプレビュー](/customization/live-preview) ボタンをクリックして、新しいブラウザタブでカスタムサインインページをテストできます。
